+++ src/sdk/account/utils/getVersion.test.ts
@@ -7,19 +7,18 @@ import {
 } from "viem"
 import { afterAll, beforeAll, describe, expect, test } from "vitest"
 import { toNetwork } from "../../../test/testSetup"
-import {
-  getTestAccount,
-  killNetwork,
-  toTestClient
-} from "../../../test/testUtils"
+import { killNetwork } from "../../../test/testUtils"
 import type { NetworkConfig } from "../../../test/testUtils"
+import { createSmartAccountClient } from "../../clients/createBicoBundlerClient"
 import {
   BICONOMY_ATTESTER_ADDRESS,
   MAINNET_ADDRESS_K1_VALIDATOR_FACTORY_ADDRESS,
-  MAINNET_ADDRESS_K1_VALIDATOR_FACTORY_ADDRESS_UNTIL_0_2
+  MAINNET_ADDRESS_K1_VALIDATOR_FACTORY_ADDRESS_UNTIL_0_2,
+  RHINESTONE_ATTESTER_ADDRESS
 } from "../../constants"
 import { BICONOMY_ATTESTER_ADDRESS_UNTIL_0_1 } from "../../constants"
 import { toNexusAccount } from "../toNexusAccount"
+import { addressEquals } from "./Utils"
 import {
   isVersionOlder,
   semverCompare,
@@ -120,6 +119,14 @@ describe("utils.getVersion", () => {
       oldVersion: "0.2.1"
     })
 
+    // Create a nexus account from the future
+    const account_v3 = await toNexusAccount({
+      chain,
+      transport: http(),
+      signer: eoaAccount,
+      oldVersion: "3"
+    })
+
     // Create a nexus account with version 0.2.0
     const account_v0_2_0 = await toNexusAccount({
       chain,
@@ -136,61 +143,138 @@ describe("utils.getVersion", () => {
       oldVersion: "0.1.0"
     })
 
-    // // Create a nexus account with version 0.0.41
-    const account_v0_0_41 = await toNexusAccount({
+    // Create a nexus account with version 0.0.40
+    const account_v0_0_40 = await toNexusAccount({
       chain,
       transport: http(),
       signer: eoaAccount,
-      oldVersion: "0.0.41"
+      oldVersion: "0.0.40"
     })
 
-    // // Create a nexus account with version 0.0.33
-    const account_v0_0_33 = await toNexusAccount({
+    // Create a nexus account with version 0.0.40
+    const account_v0_0_45 = await toNexusAccount({
       chain,
       transport: http(),
       signer: eoaAccount,
-      oldVersion: "0.0.33"
+      oldVersion: "0.0.45"
     })
+
+    // Create a nexus account with version 0.0.32
+    const account_v0_0_32 = await toNexusAccount({
+      chain,
+      transport: http(),
+      signer: eoaAccount,
+      oldVersion: "0.0.32"
+    })
+
+    //Create a nexus account with version 0.0.40
+    const account_v0_0_1 = await toNexusAccount({
+      chain,
+      transport: http(),
+      signer: eoaAccount,
+      oldVersion: "0.0.1"
+    })
+
+    const accounts = [
+      defaultAccount,
+      account_v3,
+      account_v0_2_1,
+      account_v0_2_0,
+      account_v0_1_0,
+      account_v0_0_45,
+      account_v0_0_40,
+      account_v0_0_32,
+      account_v0_0_1
+    ]
+
+    const clients = await Promise.all(
+      accounts.map((account) =>
+        createSmartAccountClient({
+          account,
+          transport: http(bundlerUrl)
+        })
+      )
+    )
+
+    const [
+      defaultAccountId,
+      account_v3Id,
+      account_v0_2_1Id,
+      account_v0_2_0Id,
+      account_v0_1_0Id,
+      account_v0_0_45Id,
+      account_v0_0_40Id,
+      account_v0_0_32Id,
+      account_v0_0_1Id
+    ] = await Promise.all(clients.map((client) => client.accountId()))
+
+    const EXPECTED_ACCOUNT_IDS = [
+      [
+        "biconomy.nexus.1.0.2",
+        defaultAccountId,
+        account_v3Id,
+        account_v0_2_1Id,
+        account_v0_2_0Id
+      ],
+      [
+        "biconomy.nexus.1.0.0",
+        account_v0_1_0Id,
+        account_v0_0_45Id,
+        account_v0_0_40Id,
+        account_v0_0_32Id,
+        account_v0_0_1Id
+      ]
+    ]
+
+    for (const batch of EXPECTED_ACCOUNT_IDS) {
+      expect(batch.every((id) => id === batch[0])).toBeTruthy()
+    }
 
     // Verify that accounts were created successfully
     const defaultAddress = await defaultAccount.getAddress()
+    const v3Address = await account_v3.getAddress()
     const v0_2_1Address = await account_v0_2_1.getAddress()
     const v0_2_0Address = await account_v0_2_0.getAddress()
     const v0_1_0Address = await account_v0_1_0.getAddress()
-    const v0_0_41Address = await account_v0_0_41.getAddress()
-    const v0_0_33Address = await account_v0_0_33.getAddress()
+    const v0_0_45Address = await account_v0_0_45.getAddress()
+    const v0_0_40Address = await account_v0_0_40.getAddress()
+    const v0_0_32Address = await account_v0_0_32.getAddress()
+    const v0_0_1Address = await account_v0_0_1.getAddress()
 
     expect(isAddress(defaultAddress)).toBeTruthy()
+    expect(isAddress(v3Address)).toBeTruthy()
     expect(isAddress(v0_2_1Address)).toBeTruthy()
     expect(isAddress(v0_2_0Address)).toBeTruthy()
     expect(isAddress(v0_1_0Address)).toBeTruthy()
-    expect(isAddress(v0_0_41Address)).toBeTruthy()
-    expect(isAddress(v0_0_33Address)).toBeTruthy()
+    expect(isAddress(v0_0_45Address)).toBeTruthy()
+    expect(isAddress(v0_0_40Address)).toBeTruthy()
+    expect(isAddress(v0_0_32Address)).toBeTruthy()
+    expect(isAddress(v0_0_1Address)).toBeTruthy()
 
-    // Uncomment these when the factory address is updated...
+    const EXPECTED_SIMILAR_ADDRESSES = [
+      [defaultAddress, v3Address, v0_2_1Address, v0_2_0Address],
+      [v0_1_0Address],
+      [v0_0_1Address, v0_0_32Address],
+      [v0_0_45Address, v0_0_40Address]
+    ]
 
-    // expect(defaultAddress).toBe(v0_2_1Address)
-    // expect(defaultAddress).toBe(v0_2_0Address)
-    // expect(defaultAddress).not.toBe(v0_1_0Address)
-    // expect(defaultAddress).not.toBe(v0_0_41Address)
-    // expect(defaultAddress).not.toBe(v0_0_33Address)
-
-    // expect(v0_2_1Address).toBe(v0_2_0Address)
-    // expect(v0_2_1Address).not.toBe(v0_1_0Address)
-    // expect(v0_2_1Address).not.toBe(v0_0_41Address)
-    // expect(v0_2_1Address).not.toBe(v0_0_33Address)
+    for (const batch of EXPECTED_SIMILAR_ADDRESSES) {
+      expect(
+        batch.every((address) => addressEquals(address, batch[0]))
+      ).toBeTruthy()
+    }
 
     // Verify that the correct attester addresses were used based on version
     expect(defaultAccount.attesters[1]).toBe(BICONOMY_ATTESTER_ADDRESS)
     expect(account_v0_2_1.attesters[1]).toBe(BICONOMY_ATTESTER_ADDRESS)
     expect(account_v0_2_0.attesters[1]).toBe(BICONOMY_ATTESTER_ADDRESS)
     expect(account_v0_1_0.attesters[1]).toBe(BICONOMY_ATTESTER_ADDRESS)
-    expect(account_v0_0_41.attesters[1]).toBe(
+
+    expect(account_v0_0_40.attesters[1]).toBe(
       BICONOMY_ATTESTER_ADDRESS_UNTIL_0_1
     )
-    expect(account_v0_0_33.attesters[1]).toBe(
-      BICONOMY_ATTESTER_ADDRESS_UNTIL_0_1
-    )
+    expect(account_v0_0_32.attesters[0]).toBe(RHINESTONE_ATTESTER_ADDRESS)
+    expect(account_v0_0_32.attesters).toHaveLength(1)
 
     expect(defaultAccount.factoryAddress).toBe(
       MAINNET_ADDRESS_K1_VALIDATOR_FACTORY_ADDRESS
@@ -204,10 +288,10 @@ describe("utils.getVersion", () => {
     expect(account_v0_1_0.factoryAddress).toBe(
       MAINNET_ADDRESS_K1_VALIDATOR_FACTORY_ADDRESS_UNTIL_0_2
     )
-    expect(account_v0_0_41.factoryAddress).toBe(
+    expect(account_v0_0_40.factoryAddress).toBe(
       MAINNET_ADDRESS_K1_VALIDATOR_FACTORY_ADDRESS_UNTIL_0_2
     )
-    expect(account_v0_0_33.factoryAddress).toBe(
+    expect(account_v0_0_32.factoryAddress).toBe(
       MAINNET_ADDRESS_K1_VALIDATOR_FACTORY_ADDRESS_UNTIL_0_2
     )
   })