+++ src/sdk/constants/index.ts
@@ -7,6 +7,10 @@ import {
   getValueLimitPolicy
 } from "@rhinestone/module-sdk"
 import { type Hex, toBytes, toHex } from "viem"
+import {
+  DEFAULT_BICONOMY_IMPLEMENTATION_ADDRESS,
+  DEFAULT_BICONOMY_IMPLEMENTATION_ADDRESS_UNTIL_0_2
+} from "../account/utils/Constants"
 import type { AddressConfig } from "../account/utils/getVersion"
 import { ParamCondition } from "../modules/smartSessionsValidator/Types"
 
@@ -56,25 +60,47 @@ export const LATEST_DEFAULT_ADDRESSES: AddressConfig = {
   attesters: [RHINESTONE_ATTESTER_ADDRESS, BICONOMY_ATTESTER_ADDRESS],
   factoryAddress: MAINNET_ADDRESS_K1_VALIDATOR_FACTORY_ADDRESS,
   bootStrapAddress: NEXUS_BOOTSTRAP_ADDRESS,
-  validatorAddress: MAINNET_ADDRESS_K1_VALIDATOR_ADDRESS
+  validatorAddress: MAINNET_ADDRESS_K1_VALIDATOR_ADDRESS,
+  accountId: "biconomy.nexus.1.0.2",
+  implementationAddress: DEFAULT_BICONOMY_IMPLEMENTATION_ADDRESS
+}
+export const EARLIEST_DEFAULT_ADDRESSES: AddressConfig = {
+  attesters: [RHINESTONE_ATTESTER_ADDRESS],
+  factoryAddress: MAINNET_ADDRESS_K1_VALIDATOR_FACTORY_ADDRESS_UNTIL_0_2,
+  bootStrapAddress: NEXUS_BOOTSTRAP_ADDRESS,
+  validatorAddress: MAINNET_ADDRESS_K1_VALIDATOR_ADDRESS,
+  accountId: "biconomy.nexus.1.0.0",
+  implementationAddress: DEFAULT_BICONOMY_IMPLEMENTATION_ADDRESS_UNTIL_0_2
 }
 
 export const DEFAULT_CONFIGURATIONS_BY_VERSION: Record<string, AddressConfig> =
   {
     "0.0": {
+      attesters: [RHINESTONE_ATTESTER_ADDRESS],
+      factoryAddress: MAINNET_ADDRESS_K1_VALIDATOR_FACTORY_ADDRESS_UNTIL_0_2,
+      bootStrapAddress: NEXUS_BOOTSTRAP_ADDRESS,
+      validatorAddress: MAINNET_ADDRESS_K1_VALIDATOR_ADDRESS,
+      accountId: "biconomy.nexus.1.0.0",
+      implementationAddress: DEFAULT_BICONOMY_IMPLEMENTATION_ADDRESS_UNTIL_0_2
+    },
+    "0.0.34": {
       attesters: [
         RHINESTONE_ATTESTER_ADDRESS,
         BICONOMY_ATTESTER_ADDRESS_UNTIL_0_1
       ],
       factoryAddress: MAINNET_ADDRESS_K1_VALIDATOR_FACTORY_ADDRESS_UNTIL_0_2,
       bootStrapAddress: NEXUS_BOOTSTRAP_ADDRESS,
-      validatorAddress: MAINNET_ADDRESS_K1_VALIDATOR_ADDRESS
+      validatorAddress: MAINNET_ADDRESS_K1_VALIDATOR_ADDRESS,
+      accountId: "biconomy.nexus.1.0.0",
+      implementationAddress: DEFAULT_BICONOMY_IMPLEMENTATION_ADDRESS_UNTIL_0_2
     },
     "0.1": {
       attesters: [RHINESTONE_ATTESTER_ADDRESS, BICONOMY_ATTESTER_ADDRESS],
       factoryAddress: MAINNET_ADDRESS_K1_VALIDATOR_FACTORY_ADDRESS_UNTIL_0_2,
       bootStrapAddress: NEXUS_BOOTSTRAP_ADDRESS,
-      validatorAddress: MAINNET_ADDRESS_K1_VALIDATOR_ADDRESS
+      validatorAddress: MAINNET_ADDRESS_K1_VALIDATOR_ADDRESS,
+      accountId: "biconomy.nexus.1.0.0",
+      implementationAddress: DEFAULT_BICONOMY_IMPLEMENTATION_ADDRESS_UNTIL_0_2
     },
     "0.2": LATEST_DEFAULT_ADDRESSES
   }